the balls on this guy

+++ src/routes/killswitches.ts
@@ -1,11 +1,15 @@
 import { logger, app } from '../index.js';
 import { Errors } from '../utils/TemplateResponses.js';
-import SessionChecker from '../middleware/SessionChecker.js';
+import ReverifyIdentity from '../middleware/ReverifyIdentity.js';
 import {
   KillSwitches,
   setKillSwitch,
   getKillSwitches,
 } from '../utils/GlobalKillSwitches.js';
+import SessionPermissionChecker, {
+  PermissionFlags,
+} from '../middleware/SessionPermissionChecker.js';
+import SessionChecker from '../middleware/SessionChecker.js';
 
 import { Request, Response } from 'express';
 
@@ -15,6 +19,7 @@ app.get(
   // GET /api/killswitches
   '/api/killswitches',
   SessionChecker(true),
+  SessionPermissionChecker(PermissionFlags.STAFF_MODIFY_KILLSWITCHES),
   async (
     req,
     res: Response<
@@ -22,6 +27,7 @@ app.get(
       | Cumulonimbus.Structures.Error
     >,
   ) => {
+    if (!req.user) return res.status(401).json(new Errors.InvalidSession());
     try {
       const killSwitches = await getKillSwitches();
 
@@ -43,7 +49,8 @@ app.get(
 app.put(
   // PUT /api/killswitches/:id
   '/api/killswitches/:id(\\d+)',
-  SessionChecker(true),
+  ReverifyIdentity(true),
+  SessionPermissionChecker(PermissionFlags.STAFF_MODIFY_KILLSWITCHES),
   async (
     req: Request<{ id: string }>,
     res: Response<
@@ -51,6 +58,7 @@ app.put(
       | Cumulonimbus.Structures.Error
     >,
   ) => {
+    if (!req.user) return res.status(401).json(new Errors.InvalidSession());
     try {
       const killSwitch = Number(req.params.id);
 
@@ -74,7 +82,8 @@ app.put(
 app.delete(
   // DELETE /api/killswitches/:id
   '/api/killswitches/:id(\\d+)',
-  SessionChecker(true),
+  ReverifyIdentity(true),
+  SessionPermissionChecker(PermissionFlags.STAFF_MODIFY_KILLSWITCHES),
   async (
     req: Request<{ id: string }>,
     res: Response<
@@ -82,6 +91,7 @@ app.delete(
       | Cumulonimbus.Structures.Error
     >,
   ) => {
+    if (!req.user) return res.status(401).json(new Errors.InvalidSession());
     try {
       const killSwitch = Number(req.params.id);
 
@@ -105,7 +115,8 @@ app.delete(
 app.delete(
   // DELETE /api/killswitches
   '/api/killswitches',
-  SessionChecker(true),
+  ReverifyIdentity(true),
+  SessionPermissionChecker(PermissionFlags.STAFF_MODIFY_KILLSWITCHES),
   async (
     req,
     res: Response<
@@ -113,6 +124,7 @@ app.delete(
       | Cumulonimbus.Structures.Error
     >,
   ) => {
+    if (!req.user) return res.status(401).json(new Errors.InvalidSession());
     try {
       logger.info(
         `User ${req.user.username} (${req.user.id}) disabled all kill switches`,
